extract page-creators from gatsby-node.js
